dz7 (event Notifications)/src/order/requirements.txt
flask
psycopg2-binary
yandex-pgmigrate
prometheus_flask_exporter
prometheus_client
gunicorn
requests
pika